Moves existing ContentBlock code to a concern to allow sufia-based apps to easily extend it.
